requirements.txt
anthropic>=0.25.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
gspread>=6.1.0
google-auth>=2.28.0
playwright>=1.43.0
schedule>=1.2.0
pdfplumber>=0.10.0
python-dotenv>=1.0.0
